renamed input to use smaller case (from temporary name)
